requirements.txt
pydantic>=2.7
sqlmodel>=0.0.16
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
openai>=1.50
instructor>=1.5
faiss-cpu>=1.8.0.post1
python-dotenv>=1.0
orjson>=3.10
pytest>=8.0
rich>=13.7
loguru>=0.7
